refactor: useAdmin index.ts에 추가
